requirements.txt
altgraph==0.17.4
antiorm==1.2.1
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
blinker==1.6.2
boto3==1.35.24
botocore==1.35.24
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
db==0.1.1
DingtalkChatbot==1.5.7
Django==4.2.4
docx==0.2.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-APScheduler==1.13.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.54.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.7
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.9.2
MyApplication==0.1.0
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.25.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.1
panda==0.3.1
pandas==2.0.3
pefile==2024.8.26
Pillow==10.0.0
protobuf==4.21.12
pycparser==2.21
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
PyMySQL==1.1.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.3
requests==2.31.0
s3transfer==0.10.2
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
xlrd==2.0.1
xmltodict==0.13.0
